+++ modules/ui_extra_networks_checkpoints.py
@@ -28,10 +28,11 @@ class ExtraNetworksPageCheckpoints(ui_extra_networks.ExtraNetworksPage):
             preview = v.get('preview', v['path'])
             preview_file = self.find_preview_file(os.path.join(reference_dir, preview))
             _size, mtime = modelstats.stat(preview_file)
+            name = os.path.normpath(os.path.join(reference_dir, k)).replace('\\', '/')
             yield {
                 "type": 'Model',
-                "name": os.path.join(reference_dir, k),
-                "title": os.path.join(reference_dir, k),
+                "name": name,
+                "title": name,
                 "filename": url,
                 "preview": self.find_preview(os.path.join(reference_dir, preview)),
                 "local_preview": preview_file,